+++ main/components/campus/helper_methods/campusMap.ui.ts
@@ -15,3 +15,51 @@ export const computeFloatingBottom = (
     ? FLOATING_BOTTOM_POSITIONS.PEEK
     : FLOATING_BOTTOM_POSITIONS.EXPANDED;
 };
+
+export function getFloatingUiState(params: {
+  isRouteMode: boolean;
+  selected: unknown;
+  isNavigating: boolean;
+  travelPopupVisible: boolean;
+  popupIndex: number;
+  windowHeight: number;
+}) {
+  const {
+    isRouteMode,
+    selected,
+    isNavigating,
+    travelPopupVisible,
+    popupIndex,
+    windowHeight,
+  } = params;
+
+  const hasBuildingPopup = !isRouteMode && !!selected;
+  const hasTravelPopup = isRouteMode && !isNavigating && travelPopupVisible;
+
+  const collapsedBuildingPopupHeight = Math.round(windowHeight * 0.19);
+  const collapsedTravelPopupHeight = Math.max(
+    260,
+    Math.round(windowHeight * 0.28),
+  );
+
+  let floatingBottom = 120;
+  if (hasBuildingPopup) {
+    floatingBottom = collapsedBuildingPopupHeight;
+  } else if (hasTravelPopup) {
+    floatingBottom = collapsedTravelPopupHeight;
+  }
+
+  const shouldShowCompass =
+    isNavigating || !(popupIndex > 0 && (hasBuildingPopup || hasTravelPopup));
+
+  const shouldHideFloatingButtons =
+    popupIndex > 0 && (hasBuildingPopup || hasTravelPopup);
+
+  return {
+    hasBuildingPopup,
+    hasTravelPopup,
+    floatingBottom,
+    shouldShowCompass,
+    shouldHideFloatingButtons,
+  };
+}
